refactor: extract JobListLogic and JobListView into there own files
